requirements.txt
google-adk>=0.2.0
datasets>=2.13.1
pytest>=7.4.0
rich>=13.5.2
gitpython>=3.1.32
pyyaml>=6.0
litellm>=1.0,<2.0